Add a "spoke" supplementary view
The supplementary view does not appear to apply the attributes from the  -initialLayoutAttributes... and -finalLayoutAttributes... methods. This is a POC that demonstrates that using "spokes" that are technically configured to zoom in and out when entering or leaving (respectively).

The new classes in the POC are:
SpokeAttributes - a subclass of UICollectionViewLayoutAttributes that adds 'origin' and 'endpoint' properties to the attributes.

SpokeReusableView - a subclass of UICollectionReusableView that's used as the spoke / supplementary view.

SpokeLayer - a CAShapeLayer subclass used by SpokeReusableView for modeling the spoke.